move HandleMissingInstances to its own function
